fixed overload ambiguity due to missing cast.
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@8245 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/VISUAL/Spectrum2DCanvas.C
@@ -2381,7 +2381,7 @@ namespace OpenMS
           QList<QVariant> chroms_idx;
           for (vector<Size>::iterator vit = mit->second.begin(); vit != mit->second.end(); ++vit)
           {
-            chroms_idx.push_back(*vit);
+            chroms_idx.push_back((unsigned int)*vit);
           }
           a->setData(chroms_idx);
